src/requirements.txt
boto3==1.36.5
python-dotenv==1.0.1
google-cloud>=0.34.0,<0.35.0
google-cloud-bigquery>=3.29.0,<4.0.0
google-cloud-storage>=2.19.0,<3.0.0
pytest==8.3.4
aws-lambda-powertools==3.9.0
opensearch-py==2.8.0
pandas==2.2.3
db-dtypes==1.4.2